Added filter parameter to Convert RPC.
p_filtr should either be equal "". That means no filter is set.
or the filter is "<expression>". In this case only transactions with names
including the expression will be converted.

Following the Zera convem:tions using "Snapshot" as expression would end up with
only Snapshots and no records in the python conversion dict.

Signed-off-by: bhamacher <b.hamacher@zera.de>